requirements.txt
absl-py==0.13.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cvzone==1.5.2
cycler==0.10.0
dlib==19.22.1
docutils==0.17.1
face-recognition==1.3.0
face-recognition-models==0.3.0
ffpyplayer==4.3.2
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
Kivy==2.0.0
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
kiwisolver==1.3.2
matplotlib==3.4.3
mediapipe-rpi4==0.8.4.0
networkx==2.6.3
numpy==1.21.2
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
Pillow==8.3.2
pkg-resources==0.0.0
protobuf==3.18.0
pygame==2.0.1
Pygments==2.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
PyWavelets==1.1.1
requests==2.26.0
RPi.GPIO==0.7.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
SPI-Py==1.0
tifffile==2021.8.30
typing-extensions==3.10.0.2
urllib3==1.26.6
zipp==3.5.0
